# nadiki_registrar/controllers/rack_api_spec_other_controller.py
import connexion
import six
from typing import Dict
from typing import Tuple
from typing import Union

from nadiki_registrar.models.error import Error  # noqa: E501
from nadiki_registrar.models.list_racks200_response import ListRacks200Response  # noqa: E501
from nadiki_registrar.models.rack_create import RackCreate  # noqa: E501
from nadiki_registrar.models.rack_metrics_query import RackMetricsQuery  # noqa: E501
from nadiki_registrar.models.rack_metrics_query_response import RackMetricsQueryResponse  # noqa: E501
from nadiki_registrar.models.rack_response import RackResponse  # noqa: E501
from nadiki_registrar.models.rack_update import RackUpdate  # noqa: E501
from nadiki_registrar import util


def create_rack(rack_create):  # noqa: E501
    """Register a new rack

    Create a new rack entry in the registry # noqa: E501

    :param rack_create: 
    :type rack_create: dict | bytes

    :rtype: Union[RackResponse, Tuple[RackResponse, int], Tuple[RackResponse, int, Dict[str, str]]
    """
    if connexion.request.is_json:
        rack_create = RackCreate.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def delete_rack(rack_id):  # noqa: E501
    """Delete rack

    Remove a rack from the registry # noqa: E501

    :param rack_id: Unique rack identifier (format RACK-[FACILITY_ID]-[RACK_ID])
    :type rack_id: str

    :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]
    """
    return 'do some magic!'


def get_rack(rack_id):  # noqa: E501
    """Get rack details

    Retrieve detailed information about a specific rack # noqa: E501

    :param rack_id: Unique rack identifier (format RACK-[FACILITY_ID]-[RACK_ID])
    :type rack_id: str

    :rtype: Union[RackResponse, Tuple[RackResponse, int], Tuple[RackResponse, int, Dict[str, str]]
    """
    return 'do some magic!'


def list_racks(limit=None, offset=None, facility_id=None):  # noqa: E501
    """List all racks

    Retrieve a list of all registered racks # noqa: E501

    :param limit: Maximum number of racks to return
    :type limit: int
    :param offset: Number of racks to skip
    :type offset: int
    :param facility_id: Filter racks by facility ID
    :type facility_id: str

    :rtype: Union[ListRacks200Response, Tuple[ListRacks200Response, int], Tuple[ListRacks200Response, int, Dict[str, str]]
    """
    return 'do some magic!'


def query_rack_metrics(rack_id, rack_metrics_query):  # noqa: E501
    """Query rack metrics

    Retrieve aggregated time series metrics for a rack over a specified time period # noqa: E501

    :param rack_id: Unique rack identifier (format RACK-[FACILITY_ID]-[RACK_ID])
    :type rack_id: str
    :param rack_metrics_query: 
    :type rack_metrics_query: dict | bytes

    :rtype: Union[RackMetricsQueryResponse, Tuple[RackMetricsQueryResponse, int], Tuple[RackMetricsQueryResponse, int, Dict[str, str]]
    """
    if connexion.request.is_json:
        rack_metrics_query = RackMetricsQuery.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def update_rack(rack_id, rack_update):  # noqa: E501
    """Update rack

    Update all rack information # noqa: E501

    :param rack_id: Unique rack identifier (format RACK-[FACILITY_ID]-[RACK_ID])
    :type rack_id: str
    :param rack_update: 
    :type rack_update: dict | bytes

    :rtype: Union[RackResponse, Tuple[RackResponse, int], Tuple[RackResponse, int, Dict[str, str]]
    """
    if connexion.request.is_json:
        rack_update = RackUpdate.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'