# nadiki_registrar/openapi/openapi.yaml
openapi: 3.0.0
info:
  description: API for registering and managing data center facilities
  title: Facility Registry API
  version: 1.0.0
servers:
- description: Registry API v1
  url: https://api.registry.example.com/v1
...
paths:
  /facilities:
    get:
      description: Retrieve a list of all registered facilities
      operationId: list_facilities
      parameters:
      - description: Maximum number of facilities to return
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          type: integer
        style: form
      - description: Number of facilities to skip
        explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listFacilities_200_response'
          description: List of facilities
      summary: List all facilities
      tags:
      - facility-api.spec_other
      x-openapi-router-controller: nadiki_registrar.controllers.facility_api_spec_other_controller
    post:
      description: Create a new facility entry in the registry
      operationId: create_facility
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FacilityCreate'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacilityResponse'
          description: Facility successfully created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid input
      summary: Register a new facility
      tags:
      - facility-api.spec_other
      x-openapi-router-controller: nadiki_registrar.controllers.facility_api_spec_other_controller
  /facilities/{facilityId}:
    delete:
      description: Remove a facility from the registry
      operationId: delete_facility
      parameters:
      - description: Unique facility identifier
        explode: false
        in: path
        name: facilityId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Facility successfully deleted
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Facility not found
      summary: Delete facility
      tags:
      - facility-api.spec_other
      x-openapi-router-controller: nadiki_registrar.controllers.facility_api_spec_other_controller
    get:
      description: Retrieve detailed information about a specific facility
      operationId: get_facility
      parameters:
      - description: Unique facility identifier
        explode: false
        in: path
        name: facilityId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacilityResponse'
          description: Facility details
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Facility not found
      summary: Get facility details
      tags:
      - facility-api.spec_other
      x-openapi-router-controller: nadiki_registrar.controllers.facility_api_spec_other_controller
    put:
      description: Update all facility information
      operationId: update_facility
      parameters:
      - description: Unique facility identifier
        explode: false
        in: path
        name: facilityId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FacilityUpdate'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacilityResponse'
          description: Facility successfully updated
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid input
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Facility not found
      summary: Update facility
      tags:
      - facility-api.spec_other
      x-openapi-router-controller: nadiki_registrar.controllers.facility_api_spec_other_controller
  /facilities/{facilityId}/query:
    post:
      description: Retrieve aggregated time series metrics for a facility over a specified
        time period
      operationId: query_facility_metrics
      parameters:
      - description: Unique facility identifier
        explode: false
        in: path
        name: facilityId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FacilityMetricsQuery'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacilityMetricsQueryResponse'
          description: Aggregated metrics data
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid query parameters
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Facility not found
      summary: Query facility metrics
      tags:
      - facility-api.spec_other
      x-openapi-router-controller: nadiki_registrar.controllers.facility_api_spec_other_controller
  /racks:
    get:
      description: Retrieve a list of all registered racks
      operationId: list_racks
      parameters:
      - description: Maximum number of racks to return
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          type: integer
        style: form
      - description: Number of racks to skip
        explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          type: integer
        style: form
      - description: Filter racks by facility ID
        explode: true
        in: query
        name: facility_id
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listRacks_200_response'
          description: List of racks
      summary: List all racks
      tags:
      - rack-api.spec_other
      x-openapi-router-controller: nadiki_registrar.controllers.rack_api_spec_other_controller
    post:
      description: Create a new rack entry in the registry
      operationId: create_rack
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RackCreate'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RackResponse'
          description: Rack successfully created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid input
      summary: Register a new rack
      tags:
      - rack-api.spec_other
      x-openapi-router-controller: nadiki_registrar.controllers.rack_api_spec_other_controller
  /racks/{rackId}:
    delete:
      description: Remove a rack from the registry
      operationId: delete_rack
      parameters:
      - description: "Unique rack identifier (format RACK-[FACILITY_ID]-[RACK_ID])"
        explode: false
        in: path
        name: rackId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Rack successfully deleted
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Rack not found
      summary: Delete rack
      tags:
      - rack-api.spec_other
      x-openapi-router-controller: nadiki_registrar.controllers.rack_api_spec_other_controller
    get:
      description: Retrieve detailed information about a specific rack
      operationId: get_rack
      parameters:
      - description: "Unique rack identifier (format RACK-[FACILITY_ID]-[RACK_ID])"
        explode: false
        in: path
        name: rackId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RackResponse'
          description: Rack details
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Rack not found
      summary: Get rack details
      tags:
      - rack-api.spec_other
      x-openapi-router-controller: nadiki_registrar.controllers.rack_api_spec_other_controller
    put:
      description: Update all rack information
      operationId: update_rack
      parameters:
      - description: "Unique rack identifier (format RACK-[FACILITY_ID]-[RACK_ID])"
        explode: false
        in: path
        name: rackId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RackUpdate'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RackResponse'
          description: Rack successfully updated
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid input
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Rack not found
      summary: Update rack
      tags:
      - rack-api.spec_other
      x-openapi-router-controller: nadiki_registrar.controllers.rack_api_spec_other_controller
  /racks/{rackId}/query:
    post:
      description: Retrieve aggregated time series metrics for a rack over a specified
        time period
      operationId: query_rack_metrics
      parameters:
      - description: "Unique rack identifier (format RACK-[FACILITY_ID]-[RACK_ID])"
        explode: false
        in: path
        name: rackId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RackMetricsQuery'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RackMetricsQueryResponse'
          description: Aggregated metrics data
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid query parameters
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Rack not found
      summary: Query rack metrics
      tags:
      - rack-api.spec_other
      x-openapi-router-controller: nadiki_registrar.controllers.rack_api_spec_other_controller
...
components:
  schemas:
    FacilityMetricsQuery:
      example:
        startTime: 2000-01-23T04:56:07.000+00:00
        aggregation: sum
        endTime: 2000-01-23T04:56:07.000+00:00
        metrics:
        - heatpump_power_consumption_joules
        - heatpump_power_consumption_joules
      properties:
        startTime:
          description: Start timestamp for the query period
          format: date-time
          title: startTime
          type: string
        endTime:
          description: End timestamp for the query period
          format: date-time
          title: endTime
          type: string
        aggregation:
          description: Aggregation function to apply over the time period
          enum:
          - sum
          - average
          title: aggregation
          type: string
        metrics:
          description: "Optional list of specific metrics to query. If not provided,\
            \ all available metrics will be returned."
          items:
            enum:
            - heatpump_power_consumption_joules
            - office_energy_use_joules
            - dc_water_usage_cubic_meters
            - office_water_usage_cubic_meters
            - total_generator_energy_joules
            - generator_load_factor_ratio
            - grid_transformers_energy_joules
            - onsite_renewable_energy_joules
            - it_power_usage_level1_joules
            - it_power_usage_level2_joules
            - renewable_energy_certificates_joules
            - grid_emission_factor_grams
            - backup_emission_factor_grams
            - electricity_source
            - pue_1_ratio
            - pue_2_ratio
            type: string
          title: metrics
          type: array
      required:
      - aggregation
      - endTime
      - startTime
      title: FacilityMetricsQuery
      type: object
    FacilityMetricsQueryResponse:
      example:
        facility_id: facility_id
        startTime: 2000-01-23T04:56:07.000+00:00
        aggregation: sum
        endTime: 2000-01-23T04:56:07.000+00:00
        metrics:
        - unit: Time
          granularitySeconds: 6
          name: name
          value: 0.8008282
        - unit: Time
          granularitySeconds: 6
          name: name
          value: 0.8008282
      properties:
        facility_id:
          description: Facility identifier
          title: facility_id
          type: string
        startTime:
          description: Start timestamp of the query period
          format: date-time
          title: startTime
          type: string
        endTime:
          description: End timestamp of the query period
          format: date-time
          title: endTime
          type: string
        aggregation:
          description: Aggregation function that was applied
          enum:
          - sum
          - average
          title: aggregation
          type: string
        metrics:
          items:
            $ref: '#/components/schemas/FacilityMetricsQueryResponse_metrics_inner'
          title: metrics
          type: array
      required:
      - aggregation
      - endTime
      - facility_id
      - metrics
      - startTime
      title: FacilityMetricsQueryResponse
      type: object
    Location:
      example:
        latitude: -75.58509
        longitude: 36.988422
      properties:
        latitude:
          description: Latitude coordinate of the facility
          format: float
          maximum: 90
          minimum: -90
          title: latitude
          type: number
        longitude:
          description: Longitude coordinate of the facility
          format: float
          maximum: 180
          minimum: -180
          title: longitude
          type: number
      required:
      - latitude
      - longitude
      title: Location
      type: object
    FacilityCreate:
      example:
        embeddedGhgEmissionsFacility: 1.4658129
        maintenanceHoursGenerator: 3.6160767
        whiteSpace: 6.846853
        designPue: 1.7386281
        coolingFluids:
        - amount: 7.0614014
          gwpFactor: 9.301444
          type: type
        - amount: 7.0614014
          gwpFactor: 9.301444
          type: type
        totalSpace: 1.4894159
        gridPowerFeeds: 1
        lifetimeFacility: 1
        whiteSpaceFloors: 1
        installedCapacity: 2.027123
        lifetimeAssets: 1
        tierLevel: 1
        location:
          latitude: -75.58509
          longitude: 36.988422
        embeddedGhgEmissionsAssets: 5.637377
      properties:
        location:
          $ref: '#/components/schemas/Location'
        embeddedGhgEmissionsFacility:
          description: Embodied carbon emissions from facility construction (CO2-eq)
          format: float
          title: embeddedGhgEmissionsFacility
          type: number
        lifetimeFacility:
          default: 15
          description: Expected lifetime of the facility in years
          minimum: 1
          title: lifetimeFacility
          type: integer
        embeddedGhgEmissionsAssets:
          description: Sum of GHG emissions embodied in assets
          format: float
          title: embeddedGhgEmissionsAssets
          type: number
        lifetimeAssets:
          default: 10
          description: Expected average lifetime of assets in years
          minimum: 1
          title: lifetimeAssets
          type: integer
        coolingFluids:
          items:
            $ref: '#/components/schemas/FacilityCreate_coolingFluids_inner'
          title: coolingFluids
          type: array
        maintenanceHoursGenerator:
          description: Annual maintenance runtime hours for generators
          format: float
          title: maintenanceHoursGenerator
          type: number
        installedCapacity:
          description: Installed/rated power capacity (watts)
          format: float
          title: installedCapacity
          type: number
        gridPowerFeeds:
          default: 3
          description: Number of physical power feeds
          minimum: 1
          title: gridPowerFeeds
          type: integer
        designPue:
          default: 1.4
          description: Design Power Usage Effectiveness
          format: float
          minimum: 1
          title: designPue
          type: number
        tierLevel:
          default: 3
          description: Certified/rated tier level
          enum:
          - 1
          - 2
          - 3
          - 4
          title: tierLevel
          type: integer
        whiteSpaceFloors:
          default: 1
          description: Number of floors used for white space
          minimum: 1
          title: whiteSpaceFloors
          type: integer
        totalSpace:
          description: Total facility space (m2)
          format: float
          title: totalSpace
          type: number
        whiteSpace:
          description: Total white space (m2)
          format: float
          title: whiteSpace
          type: number
      required:
      - location
      title: FacilityCreate
      type: object
    FacilityUpdate:
      allOf:
      - $ref: '#/components/schemas/FacilityCreate'
      - required:
        - location
        type: object
      title: FacilityUpdate
...
    FacilityResponse:
      allOf:
      - $ref: '#/components/schemas/FacilityCreate'
      - $ref: '#/components/schemas/FacilityResponse_allOf'
      title: FacilityResponse
    Error:
      properties:
        code:
          description: Error code
          title: code
          type: string
        message:
          description: Error message
          title: message
          type: string
        details:
          description: Additional error details
          title: details
          type: object
      required:
      - code
      - message
      title: Error
      type: object
    RackCreate:
      example:
        total_available_cooling_capacity: 6.0274563
        total_available_power: 0.8008282
        number_of_pdus: 1
        product_passport: "{}"
        power_redundancy: 1
        facility_id: facility_id
      properties:
        facility_id:
          description: ID of the facility where the rack is located
          title: facility_id
          type: string
        total_available_power:
          default: 5
          description: Total available power in kW
          format: float
          title: total_available_power
          type: number
        total_available_cooling_capacity:
          default: 5
          description: Total available cooling capacity in kW
          format: float
          title: total_available_cooling_capacity
          type: number
        number_of_pdus:
          default: 2
          description: Number of PDUs in the rack
          minimum: 1
          title: number_of_pdus
          type: integer
        power_redundancy:
          default: 2
          description: Number of power feeds used for redundancy
          minimum: 1
          title: power_redundancy
          type: integer
        product_passport:
          description: LCA product passport data
          title: product_passport
          type: object
      required:
      - facility_id
      title: RackCreate
      type: object
    RackUpdate:
      allOf:
      - $ref: '#/components/schemas/RackCreate'
      - required:
        - facility_id
        type: object
      title: RackUpdate
...
    RackResponse:
      allOf:
      - $ref: '#/components/schemas/RackCreate'
      - $ref: '#/components/schemas/RackResponse_allOf'
      title: RackResponse
    RackMetricsQuery:
      example:
        startTime: 2000-01-23T04:56:07.000+00:00
        aggregation: sum
        endTime: 2000-01-23T04:56:07.000+00:00
        metrics:
        - metrics
        - metrics
      properties:
        startTime:
          description: Start timestamp for the query period
          format: date-time
          title: startTime
          type: string
        endTime:
          description: End timestamp for the query period
          format: date-time
          title: endTime
          type: string
        aggregation:
          description: Aggregation function to apply over the time period
          enum:
          - sum
          - average
          title: aggregation
          type: string
        metrics:
          description: "Optional list of specific metrics to query. If not provided,\
            \ all available metrics will be returned."
          items:
            pattern: "^(pdu_[1-9][0-9]*_energy_consumption_joules|inlet_temperature_celsius|outlet_temperature_celsius)$"
            type: string
          title: metrics
          type: array
      required:
      - aggregation
      - endTime
      - startTime
      title: RackMetricsQuery
      type: object
    RackMetricsQueryResponse:
      example:
        rack_id: rack_id
        facility_id: facility_id
        startTime: 2000-01-23T04:56:07.000+00:00
        aggregation: sum
        endTime: 2000-01-23T04:56:07.000+00:00
        metrics:
        - unit: Time
          granularitySeconds: 6
          name: name
          value: 0.8008282
        - unit: Time
          granularitySeconds: 6
          name: name
          value: 0.8008282
      properties:
        facility_id:
          description: Facility identifier
          title: facility_id
          type: string
        rack_id:
          description: "Rack identifier (format RACK-[FACILITY_ID]-[RACK_ID])"
          title: rack_id
          type: string
        startTime:
          description: Start timestamp of the query period
          format: date-time
          title: startTime
          type: string
        endTime:
          description: End timestamp of the query period
          format: date-time
          title: endTime
          type: string
        aggregation:
          description: Aggregation function that was applied
          enum:
          - sum
          - average
          title: aggregation
          type: string
        metrics:
          items:
            $ref: '#/components/schemas/RackMetricsQueryResponse_metrics_inner'
          title: metrics
          type: array
      required:
      - aggregation
      - endTime
      - facility_id
      - metrics
      - rack_id
      - startTime
      title: RackMetricsQueryResponse
      type: object
...
    listFacilities_200_response:
      example:
        total: 0
        offset: 1
        limit: 6
        items:
        - null
        - null
      properties:
        items:
          items:
            $ref: '#/components/schemas/FacilityResponse'
          title: items
          type: array
        total:
          description: Total number of facilities
          title: total
          type: integer
        limit:
          description: Number of facilities per page
          title: limit
          type: integer
        offset:
          description: Current offset
          title: offset
          type: integer
      title: listFacilities_200_response
      type: object
    listRacks_200_response:
      example:
        total: 0
        offset: 1
        limit: 6
        items:
        - null
        - null
      properties:
        items:
          items:
            $ref: '#/components/schemas/RackResponse'
          title: items
          type: array
        total:
          description: Total number of racks
          title: total
          type: integer
        limit:
          description: Number of racks per page
          title: limit
          type: integer
        offset:
          description: Current offset
          title: offset
          type: integer
      title: listRacks_200_response
      type: object
...
    FacilityCreate_coolingFluids_inner:
      example:
        amount: 7.0614014
        gwpFactor: 9.301444
        type: type
      properties:
        type:
          description: Type identifier for the cooling fluid
          title: type
          type: string
        amount:
          description: Amount of cooling fluid (kg or m3)
          format: float
          title: amount
          type: number
        gwpFactor:
          description: Global Warming Potential factor for this fluid type
          format: float
          title: gwpFactor
          type: number
      required:
      - amount
      - type
      title: FacilityCreate_coolingFluids_inner
      type: object
...
    FacilityResponse_allOf:
      properties:
        id:
          description: "Unique facility identifier (format FACILITY-[COUNTRY_CODE]-ID)"
          example: FACILITY-NLD-001
          title: id
          type: string
        countryCode:
          description: ISO 3166-1 alpha-3 country code derived from location
          example: NLD
          title: countryCode
          type: string
        timeSeriesConfig:
          $ref: '#/components/schemas/FacilityTimeSeriesConfig'
        createdAt:
          format: date-time
          title: createdAt
          type: string
        updatedAt:
          format: date-time
          title: updatedAt
          type: string
      required:
      - countryCode
      - id
      - location
      - timeSeriesConfig
      title: FacilityResponse_allOf
      type: object
      example: null
...